Improved methods for transactions and email verification added

+++ src/main/java/com/api/financeapp/dtos/RecurringTransactionDTO.java
@@ -8,5 +8,5 @@ import lombok.Setter;
 public class RecurringTransactionDTO extends TransactionDTO{
     private Integer frequency;
     private String frequencyUnit;
-    private Float monthlyBudget;
+    private Double monthlyBudget;
 }
